consolidate client interfaces, types, and enums into types/
