+++ src/client/Editor/Pages/CaseSummary.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from 'react';
 import { useParams } from 'react-router-dom';
 import CaseSummaryContent from '../Components/CaseSummaryContent';
 import { CaseSummaryDetails } from '../../../common/interfaces/caseInterface';
-import { CaseSummaryParams } from '../../Common/types/CaseSummaryParams';
+import { CaseSummaryParams } from '../../types/CaseSummaryParams';
 import AsyncContent from '../../Common/components/AsyncContent';
 import { useAsyncRequestWithTwoParams } from '../../Common/hooks/useAsyncRequest';
 import { getCaseSummary } from '../../api/NodeApi';
