+++ src/client/Supervisor/Pages/Allocate.tsx
@@ -1,11 +1,11 @@
 import { ReactElement, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { ONSPanel } from 'blaise-design-system-react-components';
-import UserRole from '../../Common/enums/UserTypes';
+import UserRole from '../../types/UserTypes';
 import ErrorPanel from '../../Common/components/ErrorPanel';
 import SuccessPanel from '../../Common/components/SuccessPanel';
 import AllocateContent from '../Components/AllocateContent';
-import { Message } from '../../Common/types/MessageType';
+import { Message } from '../../types/MessageType';
 import questionnaireDisplayName from '../../Common/functions/QuestionnaireFunctions';
 import Breadcrumbs from '../../Common/components/Breadcrumbs';
