+++ src/client/Support/Components/SupportContent.tsx
@@ -3,7 +3,7 @@ import { ReactElement } from 'react';
 import { ONSPanel } from 'blaise-design-system-react-components';
 import { QuestionnaireDetails } from '../../../common/interfaces/surveyInterface';
 import CaseSearchForm from '../../Common/components/CaseSearchForm';
-import UserRole from '../../Common/enums/UserTypes';
+import UserRole from '../../types/UserTypes';
 
 interface RSupportContentProps {
   questionnaire: QuestionnaireDetails;
