+++ src/client/types/editorInterface.ts
