+++ src/client/types/supervisorInterface.ts
