+++ src/test/client/Common/components/EditCaseContent.test.tsx
@@ -5,7 +5,7 @@ import {
 import { BrowserRouter } from 'react-router-dom';
 import { CaseEditInformation } from 'blaise-api-node-client';
 import { getSpecificCaseEditInformation, setCaseToUpdate } from '../../../../client/api/NodeApi';
-import UserRole from '../../../../client/Common/enums/UserTypes';
+import UserRole from '../../../../client/types/UserTypes';
 import EditCaseContent from '../../../../client/Common/components/EditCaseContent';
 import { caseEditInformationMockObject1 } from '../../../server/mockObjects/CaseMockObject';
