+++ src/test/client/Mappers/editorInformationMapper.test.ts
@@ -3,7 +3,7 @@ import CaseOutcome from 'blaise-api-node-client/lib/cjs/enums/caseOutcome';
 import EditedStatus from 'blaise-api-node-client/lib/cjs/enums/editedStatus';
 import Organisation from 'blaise-api-node-client/lib/cjs/enums/organisation';
 import mapEditorInformation from '../../../client/Mappers/editorInformaitionMapper';
-import { EditorInformation } from '../../../client/Interfaces/editorInterface';
+import { EditorInformation } from '../../../client/types/editorInterface';
 
 describe('Map editor informaiton', () => {
   it('It should return a correctly mapped editor information model given all details are present', () => {
