+++ src/test/client/Mappers/supervisorInformationMapper.test.ts
@@ -1,7 +1,7 @@
 import { CaseEditInformation, CaseOutcome, EditedStatus } from 'blaise-api-node-client';
 import Organisation from 'blaise-api-node-client/lib/cjs/enums/organisation';
 import mapSupervisorInformaiton from '../../../client/Mappers/supervisorInformationMapper';
-import { SupervisorInformation } from '../../../client/Interfaces/supervisorInterface';
+import { SupervisorInformation } from '../../../client/types/supervisorInterface';
 
 describe('Map editor informaiton', () => {
   const editors = [{
