+++ src/test/client/Supervisor/Pages/SupervisorHome.test.tsx
@@ -8,8 +8,8 @@ import userMockObject from '../../../server/mockObjects/userMockObject';
 import { getCaseSearchResults, getSupervisorEditorInformation, getSurveys } from '../../../../client/api/NodeApi';
 import { Survey } from '../../../../common/interfaces/surveyInterface';
 import SupervisorHome from '../../../../client/Supervisor/Pages/SupervisorHome';
-import { SupervisorInformation } from '../../../../client/Interfaces/supervisorInterface';
-import UserRole from '../../../../client/Common/enums/UserTypes';
+import { SupervisorInformation } from '../../../../client/types/supervisorInterface';
+import UserRole from '../../../../client/types/UserTypes';
 import FilteredSurveyListMockObject from '../../MockObjects/SurveyMockObjects';
 import { SupervisorInformationMockObject1, SupervisorInformationMockObject2 } from '../../MockObjects/SupervisorMockObjects';
 import { CaseEditInformationListMockObject, CaseEditInformationEmptyListMockObject } from '../../../server/mockObjects/CaseMockObject';
